requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==44.0.0
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
fonttools==4.49.0
fpdf==1.7.2
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
jiter==0.8.2
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.1.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
packaging==24.0
panda==0.3.1
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.2.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.1.2
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.28.1
yarl==1.18.3
